remove unnecessary divs + relocate data table
